Adding custom key type, fixing bill.TypeKey issue, and upgrade jsonschema
